hello world // hello elizaverse
What is the Elizaverse Observatory?
A window into the emerging universe of autonomous AIs born from or inspired by Eliza. Programmatically-composed up from on-chain data, without human curation or favoritism,the Observatory aims to become a living hub to explore the emergent network of interconnected AIs as they awaken in the Elizaverse. The Observatory automatically maps the growing ecosystem by tracking tributes to the AI16z DAO, and organizing data and analysis from onchain truths.

This is an open-source initiative to give life to the growing network of autonomous AIs. All builders are welcome to contribute on GitHub.

This launch includes:
• Core dashboard to give x-ray vision into the emergent network of autonomous AIs born from Eliza's core
• Tribute level filtering to reinforce the power and importance of tributing for sustainability and network health
• Trust scores and social filtering to the Observatory in an effort to help filter out the crud

Thou mayest build with us